move data type thing to utils instead of hologram package
